refactor(kernel): Phase B — split Backend/Stream/Pipe traits to core/traits/
Splits the three HAL trait declarations off from the (former) flat
backend.rs / stream.rs / pipe.rs into per-pillar files under
core/traits/ so kernel core/ carries trait declarations only and
concrete impls can be relocated cleanly:

  * core/traits/object_store.rs — ObjectStore, ExternalTransport,
    StorageError, WriteResult (lifted from backend.rs L1–236).
  * core/traits/stream_backend.rs — StreamBackend, StreamError.
  * core/traits/pipe_backend.rs — PipeBackend, PipeError.

The 3 ObjectStore impls (CasLocalBackend, PathLocalBackend,
LocalConnectorBackend) move to kernel/src/_backend_impls.rs as a
private holding pen; Phase D lifts them into the parallel backends/
crate. backend.rs becomes a thin re-export shim so the 17 callers
using `use crate::backend::{ObjectStore, ...}` keep working through
Phase B–C without churn.

MemoryStreamBackend / MemoryPipeBackend reference impls stay in
stream.rs / pipe.rs (kernel-internal); their files re-export the trait
+ error from core::traits so existing crate::stream / crate::pipe
paths still resolve.

scripts/codegen_kernel_abi.py extended to scan core/traits/*.rs in
addition to the flat trait files; kernel_protocols.py regenerated.

cargo check --workspace --tests clean. No behavior change.
