+++ src/nexus/core/kernel_protocols.py
@@ -101,3 +101,25 @@ class ObjectStore(Protocol):
     def rename(self, old_path: str, new_path: str) -> Any: ...
     def copy_file(self, src_path: str, dst_path: str) -> Any: ...
     def list_dir(self, path: str) -> Any: ...
+
+
+@runtime_checkable
+class PipeBackend(Protocol):
+    def push(self, data: bytes) -> Any: ...
+    def pop(self) -> Any: ...
+    def close(self) -> None: ...
+    def is_closed(self) -> bool: ...
+    def is_empty(self) -> bool: ...
+    def size(self) -> int: ...
+    def msg_count(self) -> int: ...
+
+
+@runtime_checkable
+class StreamBackend(Protocol):
+    def push(self, data: bytes) -> Any: ...
+    def read_at(self, offset: int) -> Any: ...
+    def read_batch(self, offset: int, count: int) -> Any: ...
+    def close(self) -> None: ...
+    def is_closed(self) -> bool: ...
+    def tail_offset(self) -> int: ...
+    def msg_count(self) -> int: ...
